+++ aerich/exceptions.py
@@ -2,3 +2,9 @@ class NotSupportError(Exception):
     """
     raise when features not support
     """
+
+
+class DowngradeError(Exception):
+    """
+    raise when downgrade error
+    """